dome-key: Remove minimum OS version message
New taps were no longer possible with recent versions of Homebrew, as
their internal version APIs changed. On attempting to tap the
repository, the following error would prevent it:

    $ brew audit --tap teddywing/formulae
    Error: teddywing/formulae/dome-key: uninitialized constant OS::Mac::Version::SYMBOLS

Homebrew no longer has that identifier, and moved version numbers and
keys to a different class.

Remove the version check so that we don't depend on internal Homebrew
APIs which are liable to change at any time. Actually, knowing the
project, even the public APIs are liable to change at any time. And
since the `:lion` key hasn't existed in Homebrew for quite some time
now, it doesn't really make sense to have a programmatic check for Lion
since we have no way to check for it, and Homebrew doesn't support it
regardless.